+++ applications/jsonformer/requirements.txt
@@ -1,0 +1,3 @@
+fastapi
+pydantic
+jsonformer
